before error handling

+++ backend/routes/posts.js
@@ -34,11 +34,13 @@ router.post(
   checkAuth,
   multer({ storage: storage }).single("image"),
   (req, res, next) => {
+    console.log('user data', req);
     const url = req.protocol + '://' + req.get("host");
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -55,7 +57,7 @@ router.put(
   "/:id",
   checkAuth,
   multer({ storage: storage }).single("image"),
-   (req, res, next) => {
+  (req, res, next) => {
     let imagePath = req.body.imagePath;
     if (req.file) {
       const url = req.protocol + "://" + req.get("host");
@@ -65,12 +67,18 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    console.log("post", post);
-    Post.updateOne({ _id: req.params.id }, post)
+
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
       .then(result => {
-        res.status(200).json({ message: "Updated!" })
+        if (result.nModified > 0) {
+          res.status(200).json({ message: "Updated successful!" })
+        } else {
+          res.status(401).json({ message: "Not authorized!" })
+        }
+
       });
   });
 
@@ -108,10 +116,12 @@ router.get("", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  console.log("params", req.params.id);
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Deleted" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({ message: "Deleted successful!" })
+    } else {
+      res.status(401).json({ message: "Not authorized!" })
+    }
   });
 });
